Bug 1746952 - [devtools] Acknowledge the distinct storage resource for each target in browser_storage_dynamic_windows.js. r=nchevobbe
This test was brittle as it was assuming that only one resource was notified per storage type.
But this wasn't the case since fission or EFT!
We were lucky that the assertion were correct only against the latest resource
and not against the aggregate of the many resource (one per target).

Differential Revision: https://phabricator.services.mozilla.com/D134376

UltraBlame original commit: 6746ca879e1d31eb554a0b8e2d04382bfe724653