Bug 1743885 - Add latency telemetry for search suggestions. r=Standard8
This adds a `SEARCH_SUGGESTIONS_LATENCY_MS` keyed exponential histogram. Keys
are search engine IDs for built-in search engines and "other" for non-built-in
search engines. There are 50 buckets in the range 0s to 30s, which is the same
as the Merino latency histogram and other HTTP latency histograms. Latency is
always recorded except when the remote fetch is aborted, meaning it's recorded
on success, timeout, and error so that we can measure the true latency even on
timeout and error. That's also the same as the Merino histogram.

Differential Revision: https://phabricator.services.mozilla.com/D133830

UltraBlame original commit: d5d1f75790117e66208f5f5849cc98e6c20104bd