Bug 1746869 - Change test_ext_userScripts.js test_scriptMetaData to use a separate content page to test the userScript API script. r=robwu
By bisecting which one of the test cases defined in test_ext_userScripts.js was triggering this issue, test_scriptMetaData, I did notice
that this is the only test case in this test file to be using an iframe to create the webcontent page to test the api script on.

Using a separate content page does not trigger the issue and the xpcshell-test process does exist as expected in both
in-process and remote webextensions mode.

It may be reasonable to also separately make sure that recreating the same scenario in a real browser instance (e.g. GeckoView)
does not trigger the same kind of issue.

Differential Revision: https://phabricator.services.mozilla.com/D134289

UltraBlame original commit: 28e240a67de982455bc87760e4f184eddac7555d