Bug 1747609 - Don't auto-hide panel on click inside a menupopup. r=Gijs
This matches the current macOS behavior with native context menus.

The issue is that closing the panels causes the destruction of the
menupopup while the activate event hasn't run yet, see comment 9.

Executing the command event synchronously like we do for <button>s is
probably worth it to some extent, though note that on macOS with
non-native menus it'd still need to be async because of the
blinking-before-activation that menuitems have. Gnarly :-(

Of course we always use native menus for context menus so it's not a big
deal, but it still affects the content <select> dropdown, I believe.

So this is the less-invasive change.

Differential Revision: https://phabricator.services.mozilla.com/D134777

UltraBlame original commit: 309f9ab93b03487aa1ba7e9a8c4af86546f7d42e