Bug 1747970 - Remove check for lchown. r=firefox-build-system-reviewers,mhentges
HAVE_LCHOWN is only used in nsinstall.c, for a feature we don't use.
Moreover it would need to be tested with the host compiler/headers, but
function checks are tested with the target compiler/headers so it's not
necessarily accurate. Therefore, we just remove the check.

Differential Revision: https://phabricator.services.mozilla.com/D134854

UltraBlame original commit: 290977afc97345c7e1d2fbe0d9d9c59fff24d08b