Bug 1748188 - Remove nsIPrintSettings.setupSilentPrinting. r=emilio
The background we have on why this exists is in the code comments of the code
that's being removed, and in:
https://bugzilla.mozilla.org/show_bug.cgi?id=193001#c91

It seems very unlikely that this code is necessary. It's run for silent
printing when the new UI is enabled and E10s is disabled. If it was really
necessary it seems likely we'd have seen breakage in our much more common use
cases.

Differential Revision: https://phabricator.services.mozilla.com/D134936

UltraBlame original commit: c8d66e8d587ba5ede785587a200305030a28b750