Bug 1745240 - [devtools] Move WatcherActor.isServerTargetSwitchingEnabled into session's context. r=nchevobbe,jdescottes
This boolean helps know for which BrowsingContext we should create a target or track resources.
So that it is part of what defines the context we should debug and will be handy to have
in all filtering functions we use to filter browsing context or platform objects.

Differential Revision: https://phabricator.services.mozilla.com/D134421

UltraBlame original commit: e8e4a1f1d8a105942c4b0d30f2aa6e535b43e9eb