Bug 1748845 - Add a custom mechanism to make some docs warnings fatal, r=ahal,firefox-source-docs-reviewers,sylvestre
The idea is to capture the warnings in a temporary file, and then
apply a set of regex to find any that should be treated as fatal.

This allows us to fix warnings one type at a time, and prevents us
regressing the warnings that are already fixed.

The "reference target not count" warning is added to the initial
forbidden list, so we can ensure we don't end up with internal links
pointing to nowhere.

Differential Revision: https://phabricator.services.mozilla.com/D135389

UltraBlame original commit: 6a3d6ade80e692b26776d10f122c36d0aa35d893