Bug 1748870 - Do not leave a pending JS Exception on the stack r=mgaudet
In 1746231 the JS Exception stuff was moved from the
callback call site into the callback itself.
However, when it was moved, an exception could be
added even if we were going to permit the script
execution.  This is wrong.

Move the exception stuff back to the callback site
and use the result of the callback to determine if
we add anything to the stack or not

Differential Revision: https://phabricator.services.mozilla.com/D135250

UltraBlame original commit: 8ba31aff94a2418b7c6a9bcc5669af2061c756a0