Bug 1746172 - Merge Bengali ya-phala with the preceding cluster, to prevent letter-spacing introducing a gap between the glyphs. r=TYLin
This is a pretty unique special-case; in general, Indic-script VIRAMA characters
cluster with the preceding base, but in Bengali (Bangla) script, VIRAMA combines
with a *following* YA to create the ya-phala form, which as a whole then needs to
stay associated with the preceding letter.

This works for us if the font ligates the entire letter + ya-phala cluster, but
if ya-phala remains a separate glyph then we could introduce unwanted spacing.
Checking for this case in SetupClusterBoundaries enables us to ensure that the
entire letter + ya-phala cluster stays together, regardless of the specifics of
the font implementation.

Differential Revision: https://phabricator.services.mozilla.com/D135347

UltraBlame original commit: fd2b7ad184651344dd1fb62e2fe1974a75f44e1d